+++ src/index.ts
@@ -861,10 +861,10 @@ class DevServer {
         if (ready) {
           syncFile(filename);
         } else if (!filename.endsWith('map') && !existsSync(inDest(filename))) {
-          // ignore files during initial sync if they don't exist in the target directory
+          // ignore files during initial sync if they don't exist in the target directory (except for sourcemaps)
           ignoreFiles.push(filename);
         } else {
-          this.log.debug(`Watching ${filename}`);
+          syncFile(filename);
         }
       });
       watcher.on('change', (filename: string) => {